FrameworkExam/requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
colorama==0.4.3
cryptography==2.9.2
et-xmlfile==1.0.1
glob2==0.7
idna==2.10
importlib-metadata==1.7.0
jdcal==1.4.1
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
parse==1.15.0
parse-type==0.5.2
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.3
selenium==4.0.0a6.post1
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.5
zipp==3.1.0
